42. Trapping Rain Water
Given n non-negative integers representing an elevation map where the width of each bar is 1, compute how much water it can trap after rain.
Solution: simple iterative solution. where we need to look at two main things, first the max_in_vec which describes the index of max tower/number in the vector, and we have max which describes the current max value, starting from 0 or last_idx.
I did that for easier explanation and illustration.
we start as the max index is 0 which means the max is the first element in the vector, as long as we are getting lower values we need to add the difference between the max and the current value, if we face a bigger value then this value is the new max.
we do this until the max_in_vec index.
then we do the same thing backwards from the last index to the max_in_vec index. for the same reasons its symmetrical in that sense.